sales/serializers/product.py
from rest_framework import serializers

from sales.models.product import Product


# Product Serializer
class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ('id', 'name',)
