requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==37.0.4
dj-database-url==0.5.0
Django==3.2
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
zipp==3.8.1
django-import-export==2.6.1
